Add ions for prefabs

+++ Assets/_PROJECT/ThrowGame/Scripts/Modificators/InfoThrowableObject.cs
@@ -6,5 +6,5 @@ public class InfoThrowableObject
 {
     [field: SerializeField] public Sprite Icon { get; private set; }
     [field: SerializeField][field: Min(0)] public int Damage { get; private set; }
-    [field: SerializeField][field: Range(0, 1)] public int DropChance { get; private set; }
+    [field: SerializeField][field: Range(0, 1)] public float DropChance { get; private set; }
 }
